+++ glide3x/sst1/glide3/tests/test27.c
@@ -155,7 +155,7 @@ doNothing:
     return 0;
 }
 
-static unsigned long randx = 1;
+static unsigned int randx = 1;
 
 static unsigned int iRandom (unsigned int maxr)
 {
